src/server/requirements.txt
absl-py==0.8.1
astor==0.8.0
ffmpeg-python>=0.2.0
flake8>=3.7.9
Flask>=1.1.1
Flask-CORS>=3.0.8
gast==0.2.2
google-cloud-speech>=1.2.0
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
opencv-python==3.4.5.20
opt-einsum==3.1.0
pandas>=1.0.1
pocketsphinx==0.1.15
protobuf==3.10.0
pydub==0.23.1
pytest>=5.0.0
pyOpenSSL>=19.0.0
PyYAML==5.1.2
scikit-learn==0.22.2
scipy==1.3.2
six==1.13.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
